prov/efa : extract msg ops functions from rxr_ep.c to rxr_msg.c
This patch move the msg ops and related functions from rxr_ep.c to
a separate source file rxr_msg.c.
The prefix rxr_msg was added to the function name.
