+++ prov/efa/Makefile.include
@@ -47,13 +47,15 @@ _efa_files = \
 	prov/efa/src/rxr/rxr_cq.c	\
 	prov/efa/src/rxr/rxr_ep.c	\
 	prov/efa/src/rxr/rxr_cntr.c	\
-	prov/efa/src/rxr/rxr_rma.c
+	prov/efa/src/rxr/rxr_rma.c	\
+	prov/efa/src/rxr/rxr_msg.c
 
 _efa_headers = \
 	prov/efa/src/efa.h \
 	prov/efa/src/rxr/rxr.h \
 	prov/efa/src/rxr/rxr_cntr.h \
-	prov/efa/src/rxr/rxr_rma.h
+	prov/efa/src/rxr/rxr_rma.h \
+	prov/efa/src/rxr/rxr_msg.h
 
 efa_CPPFLAGS += \
 	-I$(top_srcdir)/prov/efa/src/ \
